requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
transformers
streamlit
requests
numpy

